server/requirements.txt
Flask-HTTPAuth
flask
docker
pyyaml
pytest